test: rename puppeteer.js
