+++ tests/functional/borders-and-margins.test.js
@@ -2,7 +2,7 @@
  * @jest-environment jest-playwright-preset
  * @flow
  */
-import { screenshot } from '../utils/puppeteer.js';
+import { screenshot } from '../utils/playwright.js';
 
 it('should take in account margins and borders', async () => {
   await page.goto(`${TEST_URL}/borders-and-margins.html`);
