+++ tests/functional/computeStyles.test.js
@@ -2,7 +2,7 @@
  * @jest-environment jest-playwright-preset
  * @flow
  */
-import { scroll, screenshot } from '../utils/puppeteer.js';
+import { scroll, screenshot } from '../utils/playwright.js';
 
 // FIXME: enable them again once we migrate to Playwright
 it.skip('should grow to the left', async () => {
