+++ tests/functional/fixed-container.test.js
@@ -2,7 +2,7 @@
  * @jest-environment puppeteer
  * @flow
  */
-import { screenshot } from '../utils/puppeteer.js';
+import { screenshot } from '../utils/playwright.js';
 
 it('should position the popper on the top', async () => {
   await page.goto(`${TEST_URL}/fixed-container.html`);
