+++ tests/functional/no-common-offset-parent.test.js
@@ -2,7 +2,7 @@
  * @jest-environment jest-playwright-preset
  * @flow
  */
-import { screenshot } from '../utils/puppeteer.js';
+import { screenshot } from '../utils/playwright.js';
 
 it('should be positioned on the right', async () => {
   await page.goto(`${TEST_URL}/no-common-offset-parent.html`);
