+++ tests/functional/offset.test.js
@@ -2,7 +2,7 @@
  * @jest-environment jest-playwright-preset
  * @flow
  */
-import { screenshot } from '../utils/puppeteer.js';
+import { screenshot } from '../utils/playwright.js';
 
 it('should offset the popper correctly', async () => {
   await page.goto(`${TEST_URL}/modifiers/offset.html`);
