+++ tests/functional/relative-body.test.js
@@ -2,7 +2,7 @@
  * @jest-environment jest-playwright-preset
  * @flow
  */
-import { screenshot } from '../utils/puppeteer.js';
+import { screenshot } from '../utils/playwright.js';
 
 it('should position the popper on the right', async () => {
   await page.goto(`${TEST_URL}/relative-body.html`);
