+++ tests/functional/resize.test.js
@@ -2,7 +2,7 @@
  * @jest-environment jest-playwright-preset
  * @flow
  */
-import { screenshot } from '../utils/puppeteer.js';
+import { screenshot } from '../utils/playwright.js';
 
 it('should update the position when window is resized', async () => {
   await page.goto(`${TEST_URL}/resize.html`);
