+++ tests/functional/virtual.test.js
@@ -2,7 +2,7 @@
  * @jest-environment jest-playwright-preset
  * @flow
  */
-import { scroll, screenshot } from '../utils/puppeteer.js';
+import { scroll, screenshot } from '../utils/playwright.js';
 
 it('should follow the cursor', async () => {
   await page.goto(`${TEST_URL}/virtual.html`);
